+++ lang/attribute/program/src/declare_program/mod.rs
@@ -1,7 +1,7 @@
 mod common;
 mod mods;
 
-use anchor_lang_idl::types::Idl;
+use anchor_lang_idl::{convert::convert_idl, types::Idl};
 use anyhow::anyhow;
 use quote::{quote, ToTokens};
 use syn::parse::{Parse, ParseStream};
@@ -45,7 +45,7 @@ fn get_idl(name: &syn::Ident) -> anyhow::Result<Idl> {
         .map(|idl_dir| idl_dir.join(name.to_string()).with_extension("json"))
         .map(std::fs::read)?
         .map_err(|e| anyhow!("Failed to read IDL `{name}`: {e}"))
-        .map(|buf| Idl::from_slice_with_conversion(&buf))?
+        .map(|buf| convert_idl(&buf))?
 }
 
 fn gen_program(idl: &Idl, name: &syn::Ident) -> proc_macro2::TokenStream {
@@ -114,8 +114,12 @@ fn gen_id(idl: &Idl) -> proc_macro2::TokenStream {
         #[doc = #doc]
         pub static ID: Pubkey = __ID;
 
+        /// Const version of `ID`
+        pub const ID_CONST: Pubkey = __ID_CONST;
+
         /// The name is intentionally prefixed with `__` in order to reduce to possibility of name
         /// clashes with the crate's `ID`.
         static __ID: Pubkey = Pubkey::new_from_array([#(#address_bytes,)*]);
+        const __ID_CONST : Pubkey = Pubkey::new_from_array([#(#address_bytes,)*]);
     }
 }